# docs/swagger.yaml
definitions:
  apperrs.CustomError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  wallet.Err:
    properties:
      message:
        type: string
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Create Card
        type: string
    type: object
  wallet.WalletRequest:
    properties:
      balance:
        example: 100
        maximum: 10000
        minimum: 500
        type: number
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        maxLength: 255
        minLength: 3
        type: string
      wallet_name:
        example: John's Wallet
        maxLength: 255
        minLength: 3
        type: string
      wallet_type:
        enum:
        - Savings
        - Credit Card
        - Crypto Wallet
        example: Credit Card
        type: string
    required:
    - balance
    - user_id
    - user_name
    - wallet_name
    - wallet_type
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/users/{id}/wallets:
    delete:
      consumes:
      - application/json
      description: Delete user wallets by user id
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrs.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrs.CustomError'
      summary: Delete user wallets
      tags:
      - wallet
    get:
      consumes:
      - application/json
      description: Get user wallets
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrs.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrs.CustomError'
      summary: Get user wallets
      tags:
      - wallet
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets
      tags:
      - wallet
  /api/v1/wallets/:
    post:
      consumes:
      - application/json
      description: Create user wallets
      parameters:
      - description: WalletRequest
        in: body
        name: WalletCreateRequest
        required: true
        schema:
          $ref: '#/definitions/wallet.WalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrs.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrs.CustomError'
      summary: Create user wallets
      tags:
      - wallet
  /api/v1/wallets/{id}:
    put:
      consumes:
      - application/json
      description: Update user wallets by wallet id
      parameters:
      - description: wallet id
        in: path
        name: id
        required: true
        type: string
      - description: WalletRequest
        in: body
        name: WalletCreateRequest
        required: true
        schema:
          $ref: '#/definitions/wallet.WalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrs.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrs.CustomError'
      summary: Update user wallets
      tags:
      - wallet
swagger: "2.0"